+++ osu.Game/Beatmaps/IBeatmapSetOnlineInfo.cs
@@ -1,8 +1,4 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
 using System;
-using Newtonsoft.Json;
 
 namespace osu.Game.Beatmaps
 {
@@ -96,44 +92,5 @@ namespace osu.Game.Beatmaps
         /// Non-null only if the track is linked to a featured artist track entry.
         /// </summary>
         public int? TrackId { get; set; }
-    }
-
-    public class BeatmapSetOnlineGenre
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-
-    public class BeatmapSetOnlineLanguage
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-
-    public class BeatmapSetOnlineCovers
-    {
-        public string CoverLowRes { get; set; }
-
-        [JsonProperty(@"cover@2x")]
-        public string Cover { get; set; }
-
-        public string CardLowRes { get; set; }
-
-        [JsonProperty(@"card@2x")]
-        public string Card { get; set; }
-
-        public string ListLowRes { get; set; }
-
-        [JsonProperty(@"list@2x")]
-        public string List { get; set; }
-    }
-
-    public class BeatmapSetOnlineAvailability
-    {
-        [JsonProperty(@"download_disabled")]
-        public bool DownloadDisabled { get; set; }
-
-        [JsonProperty(@"more_information")]
-        public string ExternalLink { get; set; }
     }
 }
